fix: rename folders
