+++ .idea/misc.xml
@@ -11,21 +11,15 @@
         <entry key="..\:/jei_s/programming/manuelCourse/android-test-master/android-test-master/CounterApplication/app/src/main/res/drawable/ic_raket.xml" value="0.403125" />
         <entry key="..\:/jei_s/programming/manuelCourse/android-test-master/android-test-master/CounterApplication/app/src/main/res/layout/activity_main.xml" value="0.3211050724637681" />
         <entry key="..\:/jei_s/programming/manuelCourse/android-test-master/android-test-master/CounterApplication/app/src/main/res/layout/activity_main_screen.xml" value="0.38697916666666665" />
-        <entry key="..\:/jei_s/programming/manuelCourse/androidCounter/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.2445" />
-        <entry key="..\:/jei_s/programming/manuelCourse/androidCounter/app/src/main/res/drawable/home_icon.xml" value="0.5661458333333333" />
-        <entry key="..\:/jei_s/programming/manuelCourse/androidCounter/app/src/main/res/drawable/ic_baseline_minimize_24.xml" value="0.5661458333333333" />
-        <entry key="..\:/jei_s/programming/manuelCourse/androidCounter/app/src/main/res/drawable/ic_launcher_background.xml" value="0.5661458333333333" />
-        <entry key="..\:/jei_s/programming/manuelCourse/androidCounter/app/src/main/res/drawable/ic_minus.xml" value="0.5661458333333333" />
-        <entry key="..\:/jei_s/programming/manuelCourse/androidCounter/app/src/main/res/drawable/ic_plus.xml" value="0.5661458333333333" />
-        <entry key="..\:/jei_s/programming/manuelCourse/androidCounter/app/src/main/res/layout/activity_home.xml" value="0.33" />
-        <entry key="..\:/jei_s/programming/manuelCourse/androidCounter/app/src/main/res/layout/activity_main.xml" value="0.4433876811594203" />
-        <entry key="..\:/jei_s/programming/manuelCourse/androidCounter/app/src/main/res/layout/fragment_loading.xml" value="0.4433876811594203" />
-        <entry key="..\:/jei_s/programming/manuelCourse/androidCounter/app/src/main/res/layout/fragment_no_counters.xml" value="0.2674772036474164" />
-        <entry key="..\:/jei_s/programming/manuelCourse/androidCounter/app/src/main/res/layout/item_counter.xml" value="0.536" />
+        <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.1445945945945946" />
+        <entry key="app/src/main/res/drawable/home_icon.xml" value="0.1445945945945946" />
+        <entry key="app/src/main/res/drawable/ic_launcher_background.xml" value="0.1445945945945946" />
+        <entry key="app/src/main/res/layout/activity_home.xml" value="0.1847972972972973" />
+        <entry key="app/src/main/res/layout/activity_main.xml" value="0.1341216216216216" />
       </map>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="Android Studio default JDK" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="11" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
