--- app/src/main/java/jeison/com/counterapplication/model/Counter.java
@@ -1,19 +1,0 @@
-package jeison.com.counterapplication.model;
-
-public class Counter {
-    private String definition;
-    private int counterTime;
-
-    public Counter(String definition, int counterTime) {
-        this.definition = definition;
-        this.counterTime = counterTime;
-    }
-
-    public String getDefinition() {
-        return definition;
-    }
-
-    public int getCounterTime() {
-        return counterTime;
-    }
-}
