+++ app/src/main/java/jeison/com/counterapplication/models/Counter.java
@@ -1,0 +1,19 @@
+package jeison.com.counterapplication.models;
+
+public class Counter {
+    private String definition;
+    private int counterTime;
+
+    public Counter(String definition, int counterTime) {
+        this.definition = definition;
+        this.counterTime = counterTime;
+    }
+
+    public String getDefinition() {
+        return definition;
+    }
+
+    public int getCounterTime() {
+        return counterTime;
+    }
+}
